Changed class ChangeLogScrollView into a ChangeLogAlertDialog inner class.
